RAM8, RAM64, PC implementations
Note that the solution for PC that the tests are looking for is
actually incorrect against the specification in the HDL file and
in the book. The book says "if... else if... else if..." and
does not specify what should happen if multiple control bits are
set. If you implement it like the pseudocode implies, you end
up with undefined behavior that fail. Instead, the solution they
are looking for should be represented as "if... if... if..." for
control bits.

I'm not bitter - it just took me an hour to figure out what
they are actually looking for.